SETI App Bug Fixes + Position File Tiling Basis
This  fixes a few minor bugs in the SETI App from the GUI to App update.

It also adds in all the hooks necessary in the Framework to enable position file based tiling. Further adds the initial versions of all necessary scripts for this tiling mode. They require testing once we have data from SETI 2.0.